requirements.txt
ultralytics==8.1.1
opencv-python
numpy
Pillow
PyYAML
streamlit